ml-worker/requirements.txt
temporalio
pydantic
numpy
pandas
scikit-learn
transformers
python-jobspy
